8.1.1 Создаст класс для работы с API и перенесет все константы в const.js
