+++ src/points-api-service.js
@@ -1,0 +1,32 @@
+import ApiService from './framework/api-service.js';
+import {Method} from './const.js';
+
+export default class PointsApiService extends ApiService {
+  get points() {
+    return this._load({url: 'points'})
+      .then(ApiService.parseResponse);
+  }
+
+  get offers() {
+    return this._load({url: 'offers'})
+      .then(ApiService.parseResponse);
+  }
+
+  get destinations() {
+    return this._load({url: 'destinations'})
+      .then(ApiService.parseResponse);
+  }
+
+  async updatePoin(point) {
+    const response = await this._load({
+      url: `points/${point.id}`,
+      method: Method.PUT,
+      body: JSON.stringify(point),
+      headers: new Headers({'Content-Type': 'application/json'}),
+    });
+
+    const parsedResponse = await ApiService.parseResponse(response);
+
+    return parsedResponse;
+  }
+}
