+++ src/view/no-route-point-view.js
@@ -1,12 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import {FilterType} from '../const.js';
-
-const NoPointsTextType = {
-  [FilterType.EVERYTHING]: 'Click New Event to create your first point',
-  [FilterType.FUTURE]: 'There are no future events now',
-  [FilterType.PRESENT]: 'There are no present events now',
-  [FilterType.PAST]: 'There are no past events now',
-};
+import {NoPointsTextType} from '../const.js';
 
 function createNoRoutePointTemplate(filterType) {
   const noPointsTextValue = NoPointsTextType[filterType];
